Service part finish and need testing

+++ app/src/main/java/tw/geodoer/mPriority/API/WeightCalculator.java
@@ -46,7 +46,7 @@ public class WeightCalculator
 
 		return Lpart + Tpart;
 	}
-	public int getweight(int pT, double pL)
+	public int getweight(long pT, double pL)
 	{
         double T = TimeUnit.MILLISECONDS.toMinutes(pT);
         double L = pL*1000;
